HUGE BUGFIX: removed "\n" character from file header
This caused a HUGE bug. The header is supposed to be only 3 bytes long, but the additional "\n" previously made it 4 bytes long. When reading the files and skipping the 3bytes-long header (while the header was actually 4bytes long), this created a one-byte offset, which resulted in the whole dataset being garbage.